+++ SnakeGame/SnakeGame/Pixel.cs
@@ -1,0 +1,51 @@
+﻿using System;
+using static System.Console;
+
+namespace SnakeGame
+{
+    public readonly struct Pixel
+    {
+        private const char PixelChar = '█';
+
+        public Pixel(int x, int y, ConsoleColor color, int pixelSize = 3)
+        {
+            X = x;
+            Y = y;
+            Color = color;
+            PixelSize = pixelSize;
+        }
+
+        public int X { get; }
+
+        public int Y { get; }
+
+        public ConsoleColor Color { get; }
+
+        public int PixelSize { get; }
+
+        public void Draw()
+        {
+            ForegroundColor = Color;
+            for (int x = 0; x < PixelSize; x++)
+            {
+                for (int y = 0; y < PixelSize; y++)
+                {
+                    SetCursorPosition(X * PixelSize + x, Y * PixelSize + y);
+                    Write(PixelChar);
+                }
+            }
+        }
+
+        public void Clear()
+        {
+            for (int x = 0; x < PixelSize; x++)
+            {
+                for (int y = 0; y < PixelSize; y++)
+                {
+                    SetCursorPosition(X * PixelSize + x, Y * PixelSize + y);
+                    Write(' ');
+                }
+            }
+        }
+    }
+}
