+++ SnakeGame/SnakeGame/Snake.cs
@@ -1,0 +1,84 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SnakeGame
+{
+    public class Snake
+    {
+        private readonly ConsoleColor _headColor;
+        private readonly ConsoleColor _bodyColor;
+
+        public Snake(int initialX,
+            int initialY,
+            ConsoleColor headColor,
+            ConsoleColor bodyColor,
+            int bodyLength = 3)
+        {
+            _headColor = headColor;
+            _bodyColor = bodyColor;
+
+            Head = new Pixel(initialX, initialY, headColor);
+
+            for (int i = bodyLength; i >= 0; i--)
+            {
+                Body.Enqueue(new Pixel(Head.X - i - 1, initialY, _bodyColor));
+            }
+
+            Draw();
+        }
+
+        public Pixel Head { get; private set; }
+        public Queue<Pixel> Body { get; } = new Queue<Pixel>();
+
+        public void Move(Direction direction, bool eat = false)
+        {
+            Clear();
+
+            Body.Enqueue(new Pixel(Head.X, Head.Y, _bodyColor));
+            if (!eat)
+                Body.Dequeue();
+
+            if (direction == Direction.Right)
+            {
+                Head = new Pixel(Head.X + 1, Head.Y, _headColor);
+            }
+            else if (direction == Direction.Left)
+            {
+                Head = new Pixel(Head.X - 1, Head.Y, _headColor);
+            }
+            else if (direction == Direction.Up)
+            {
+                Head = new Pixel(Head.X, Head.Y - 1, _headColor);
+            }
+            else if (direction == Direction.Down)
+            {
+                Head = new Pixel(Head.X, Head.Y + 1, _headColor);
+            }
+
+            Draw();
+        }
+
+        public void Draw()
+        {
+            Head.Draw();
+
+            foreach (Pixel pixel in Body)
+            {
+                pixel.Draw();
+            }
+        }
+
+        public void Clear()
+        {
+            Head.Clear();
+
+            foreach (Pixel pixel in Body)
+            {
+                pixel.Clear();
+            }
+        }
+    }
+}
