+++ SnakeGame/SnakeGame/bin/Debug/leaderboard.txt
@@ -8,3 +8,8 @@
 ааа: 0
 : 0
 : 0
+d: 0
+аа: 0
+вв: 0
+: 0
+: 0
