+++ SnakeGame/SnakeGame/bin/Debug/records.txt
@@ -1,0 +1,10 @@
+sdsd 2
+dd 2
+ 2
+dd 1
+ 1
+вв 1
+ 0
+ 0
+wd 0
+ 0
